+++ lib/src/author/runtime.rs
@@ -127,10 +127,10 @@ pub struct Success {
 #[derive(Debug, derive_more::Display)]
 pub enum Error {
     /// Error while executing the Wasm virtual machine.
-    #[display(fmt = "{}", _0)]
+    #[display(fmt = "{_0}")]
     WasmVm(runtime_host::Error),
     /// Error while initializing the Wasm virtual machine.
-    #[display(fmt = "{}", _0)]
+    #[display(fmt = "{_0}")]
     VmInit(host::StartErr, host::HostVmPrototype),
     /// Overflow when incrementing block height.
     BlockHeightOverflow,
@@ -141,11 +141,7 @@ pub enum Error {
     /// Error while parsing output of `BlockBuilder_apply_extrinsic`.
     BadApplyExtrinsicOutput,
     /// Applying an inherent extrinsic has returned a [`DispatchError`].
-    #[display(
-        fmt = "Error while applying inherent extrinsic: {}\nExtrinsic: {:?}",
-        error,
-        extrinsic
-    )]
+    #[display(fmt = "Error while applying inherent extrinsic: {error}\nExtrinsic: {extrinsic:?}")]
     InherentExtrinsicDispatchError {
         /// Extrinsic that triggered the problem.
         extrinsic: Vec<u8>,
@@ -153,11 +149,7 @@ pub enum Error {
         error: DispatchError,
     },
     /// Applying an inherent extrinsic has returned a [`TransactionValidityError`].
-    #[display(
-        fmt = "Error while applying inherent extrinsic: {}\nExtrinsic: {:?}",
-        error,
-        extrinsic
-    )]
+    #[display(fmt = "Error while applying inherent extrinsic: {error}\nExtrinsic: {extrinsic:?}")]
     InherentExtrinsicTransactionValidityError {
         /// Extrinsic that triggered the problem.
         extrinsic: Vec<u8>,
@@ -698,10 +690,10 @@ fn parse_apply_extrinsic_output(
 #[derive(Debug, derive_more::Display, Clone, PartialEq, Eq)]
 pub enum TransactionValidityError {
     /// The transaction is invalid.
-    #[display(fmt = "Transaction is invalid: {}", _0)]
+    #[display(fmt = "Transaction is invalid: {_0}")]
     Invalid(InvalidTransaction),
     /// Transaction validity can't be determined.
-    #[display(fmt = "Transaction validity couldn't be determined: {}", _0)]
+    #[display(fmt = "Transaction validity couldn't be determined: {_0}")]
     Unknown(UnknownTransaction),
 }
 
@@ -734,7 +726,7 @@ pub enum InvalidTransaction {
     /// left in the current block.
     ExhaustsResources,
     /// Any other custom invalid validity that is not covered by this enum.
-    #[display(fmt = "Other reason (code: {})", _0)]
+    #[display(fmt = "Other reason (code: {_0})")]
     Custom(u8),
     /// An extrinsic with a Mandatory dispatch resulted in Error. This is indicative of either a
     /// malicious validator or a buggy `provide_inherent`. In any case, it can result in dangerously
@@ -753,7 +745,7 @@ pub enum UnknownTransaction {
     /// No validator found for the given unsigned transaction.
     NoUnsignedValidator,
     /// Any other custom unknown validity that is not covered by this enum.
-    #[display(fmt = "Other reason (code: {})", _0)]
+    #[display(fmt = "Other reason (code: {_0})")]
     Custom(u8),
 }
 
@@ -765,7 +757,7 @@ pub enum DispatchError {
     /// A bad origin.
     BadOrigin,
     /// A custom error in a module.
-    #[display(fmt = "Error in module #{}, error number #{}", index, error)]
+    #[display(fmt = "Error in module #{index}, error number #{error}")]
     Module {
         /// Module index, matching the metadata module index.
         index: u8,