+++ lib/src/trie.rs
@@ -107,11 +107,10 @@ use core::iter;
 mod nibble;
 
 pub mod calculate_root;
-pub mod node_value;
 pub mod prefix_proof;
 pub mod proof_decode;
 pub mod proof_encode;
-pub mod proof_node_codec;
+pub mod trie_node;
 pub mod trie_structure;
 
 pub use nibble::{