+++ lib/src/trie/trie_structure.rs
@@ -389,12 +389,8 @@ impl<TUd> TrieStructure<TUd> {
                 // The trie is empty, or the key of the root node of the trie doesn't start with
                 // the requested prefix, or the key of the root node of the trie starts with the
                 // requested prefix.
-                let root_index = if let Some(i) = self.root_index {
-                    i
-                } else {
-                    // Trie is empty. Nothing to do.
-                    return None;
-                };
+                // If the trie is empty. then there is nothing to do and we return `None`.
+                let root_index = self.root_index?;
 
                 // Compare root key with the prefix.
                 if !self.nodes[root_index]
@@ -947,7 +943,7 @@ impl<'a, TUd> NodeAccess<'a, TUd> {
     }
 
     /// Returns the partial key of the node.
-    pub fn partial_key(&'_ self) -> impl ExactSizeIterator<Item = Nibble> + '_ {
+    pub fn partial_key(&'_ self) -> impl ExactSizeIterator<Item = Nibble> + Clone + '_ {
         match self {
             NodeAccess::Storage(n) => Either::Left(n.partial_key()),
             NodeAccess::Branch(n) => Either::Right(n.partial_key()),
@@ -1075,7 +1071,7 @@ impl<'a, TUd> StorageNodeAccess<'a, TUd> {
     }
 
     /// Returns the partial key of the node.
-    pub fn partial_key(&'_ self) -> impl ExactSizeIterator<Item = Nibble> + '_ {
+    pub fn partial_key(&'_ self) -> impl ExactSizeIterator<Item = Nibble> + Clone + '_ {
         self.trie
             .nodes
             .get(self.node_index)
@@ -1445,7 +1441,7 @@ impl<'a, TUd> BranchNodeAccess<'a, TUd> {
     }
 
     /// Returns the partial key of the node.
-    pub fn partial_key(&'_ self) -> impl ExactSizeIterator<Item = Nibble> + '_ {
+    pub fn partial_key(&'_ self) -> impl ExactSizeIterator<Item = Nibble> + Clone + '_ {
         self.trie
             .nodes
             .get(self.node_index)