+++ wasm-node/rust/src/lib.rs
@@ -92,9 +92,7 @@ impl Add<Duration> for Instant {
 
     fn add(self, other: Duration) -> Instant {
         let new_val = self.inner + other.as_millis() as f64;
-        Instant {
-            inner: new_val as f64,
-        }
+        Instant { inner: new_val }
     }
 }
 
@@ -103,9 +101,7 @@ impl Sub<Duration> for Instant {
 
     fn sub(self, other: Duration) -> Instant {
         let new_val = self.inner - other.as_millis() as f64;
-        Instant {
-            inner: new_val as f64,
-        }
+        Instant { inner: new_val }
     }
 }
 
@@ -257,7 +253,9 @@ fn add_chain(
                 .as_mut()
                 .unwrap()
                 .chains
-                .insert(init::Chain::Erroneous { error });
+                .insert(init::Chain::Erroneous {
+                    error: error.to_string(),
+                });
 
             return u32::try_from(chain_id).unwrap();
         }